requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
sib_api_v3_sdk==7.5.0
cryptography==41.0.7
psutil
pyotp
qrcode
